special init - wednesday
So a certain someone informed me that this is a possible test for wednesday - a certain program that determines what is the highest number, and how it reacts to inputs that are equal to each other. Read and test at your own risk. Any questions, read the readme (and FAQ.). Good evening.