init - Evolution of Text
this is probably my favorite thing to work on! I was doing some research on how some of the syntax here works and it gave me this idea. This took longer than it should have, but school work comes first. Finished around 2-3 months due to the lack of free time.